+++ src/main/java/com/telepathicgrunt/world_blender/mixin/MinecraftServerMixin.java
@@ -11,12 +11,16 @@ import com.telepathicgrunt.world_blender.the_blender.TheBlender;
 import net.minecraft.resource.ResourcePackManager;
 import net.minecraft.resource.ServerResourceManager;
 import net.minecraft.server.MinecraftServer;
+import net.minecraft.server.WorldGenerationProgressListener;
 import net.minecraft.server.WorldGenerationProgressListenerFactory;
+import net.minecraft.server.world.ServerWorld;
 import net.minecraft.structure.StructureManager;
 import net.minecraft.util.UserCache;
 import net.minecraft.util.registry.DynamicRegistryManager;
 import net.minecraft.util.registry.Registry;
+import net.minecraft.util.registry.RegistryKey;
 import net.minecraft.world.SaveProperties;
+import net.minecraft.world.World;
 import net.minecraft.world.level.storage.LevelStorage;
 import org.spongepowered.asm.mixin.Final;
 import org.spongepowered.asm.mixin.Mixin;
@@ -26,6 +30,7 @@ import org.spongepowered.asm.mixin.injection.Inject;
 import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
 
 import java.net.Proxy;
+import java.util.Map;
 
 
 @Mixin(value = MinecraftServer.class, priority = Integer.MAX_VALUE)
@@ -35,11 +40,20 @@ public class MinecraftServerMixin {
     @Shadow
     private StructureManager structureManager;
 
+    @Final
+    @Shadow
+    private Map<RegistryKey<World>, ServerWorld> worlds;
+
     @Inject(
             method = "<init>",
             at = @At(value = "TAIL")
     )
-    private void modifyBiomeRegistry(Thread thread, DynamicRegistryManager.Impl impl, LevelStorage.Session session, SaveProperties saveProperties, ResourcePackManager resourcePackManager, Proxy proxy, DataFixer dataFixer, ServerResourceManager serverResourceManager, MinecraftSessionService minecraftSessionService, GameProfileRepository gameProfileRepository, UserCache userCache, WorldGenerationProgressListenerFactory worldGenerationProgressListenerFactory, CallbackInfo ci) {
+    private void modifyBiomeRegistry(Thread thread, DynamicRegistryManager.Impl impl, LevelStorage.Session session,
+                                     SaveProperties saveProperties, ResourcePackManager resourcePackManager, Proxy proxy,
+                                     DataFixer dataFixer, ServerResourceManager serverResourceManager,
+                                     MinecraftSessionService minecraftSessionService, GameProfileRepository gameProfileRepository,
+                                     UserCache userCache, WorldGenerationProgressListenerFactory worldGenerationProgressListenerFactory,
+                                     CallbackInfo ci) {
 
         if(impl.getOptional(Registry.BIOME_KEY).isPresent()) {
             TheBlender.blendTheWorld(impl);
@@ -48,5 +62,14 @@ public class MinecraftServerMixin {
                 IdentifierPrinting.printAllIdentifiers(impl);
             }
         }
+
+    }
+
+    @Inject(
+            method = "createWorlds(Lnet/minecraft/server/WorldGenerationProgressListener;)V",
+            at = @At(value = "TAIL")
+    )
+    private void modifyBiomeRegistry(WorldGenerationProgressListener worldGenerationProgressListener, CallbackInfo ci) {
+        TheBlender.addDimensionalSpacing(worlds);
     }
 }